remove _standardize_ionq_device_properties_json and rename the test file
