+++ test/unit_tests/braket/emulation/test_local_emulator.py
@@ -22,7 +22,6 @@ from braket.device_schema.ionq.ionq_device_capabilities_v1 import IonqDeviceCapa
 from braket.device_schema.rigetti.rigetti_device_capabilities_v1 import RigettiDeviceCapabilities
 
 from braket.emulation.local_emulator import LocalEmulator
-from braket.emulation._standardization import _standardize_ionq_device_properties_json
 
 from conftest import invalid_device_properties_dict_1, invalid_device_properties_dict_2
 
@@ -89,9 +88,7 @@ def test_validate_valid_verbatim_circ_ankaa3(
 def test_validate_valid_verbatim_circ_aria_1(
     reduced_ionq_device_capabilities_json, valid_verbatim_circ_aria1
 ):
-    emulator = LocalEmulator.from_json(
-        _standardize_ionq_device_properties_json(reduced_ionq_device_capabilities_json)
-    )
+    emulator = LocalEmulator.from_json(reduced_ionq_device_capabilities_json)
     emulator.validate(valid_verbatim_circ_aria1)
